requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-cors-headers==4.6.0
pytest==8.0.0
pytest-django==4.8.0
django-filter==23.5
Pillow==11.1.0  # For image handling if needed
requests==2.32.3  # For external API calls 